--- a/src/Compiler/Checking/import.fsi
+++ b/src/Compiler/Checking/import.fsi
@@ -97,8 +97,7 @@ val internal ImportILAssembly:
 
 /// Import the type forwarder table for an IL assembly
 val internal ImportILAssemblyTypeForwarders:
-    (unit -> ImportMap) * range * ILExportedTypesAndForwarders ->
-        IImmutableDictionary<string array * string, Lazy<EntityRef>>
+    (unit -> ImportMap) * range * ILExportedTypesAndForwarders -> CcuTypeForwarderTable
 
 /// Import an IL type as an F# type, first rescoping to view the metadata from the current assembly
 /// being compiled. importInst gives the context for interpreting type variables.