--- a/src/Compiler/Driver/CompilerImports.fs
+++ b/src/Compiler/Driver/CompilerImports.fs
@@ -1115,7 +1115,7 @@ and [<Sealed>] TcImports(tcConfigP: TcConfigProvider, initialResolutions: TcAsse
                 MemberSignatureEquality = (fun ty1 ty2 -> typeEquivAux EraseAll g ty1 ty2)
                 ImportProvidedType = (fun ty -> ImportProvidedType (tcImports.GetImportMap()) m ty)
                 TryGetILModuleDef = (fun () -> Some ilModule)
-                TypeForwarders = ImmutableDictionary.Empty
+                TypeForwarders = CcuTypeForwarderTable.Empty
                 XmlDocumentationInfo =
                     match tcConfig.xmlDocInfoLoader with
                     | Some xmlDocInfoLoader -> xmlDocInfoLoader.TryLoad(fileName)