--- a/tests/benchmarks/FCSSourceFiles/FCSSourceFiles.fsproj
+++ b/tests/benchmarks/FCSSourceFiles/FCSSourceFiles.fsproj
@@ -14,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\src\fsharp\FSharp.Compiler.Service\FSharp.Compiler.Service.fsproj" />
+    <ProjectReference Include="..\..\..\src\Compiler\FSharp.Compiler.Service.fsproj" />
   </ItemGroup>
 
 </Project>