This is a naive attempt at addressing https://github.com/dotnet/fsharp/issues/13033.
I am most likely missing the point 🤔as I didn't notice much difference after changing the Map to a Dictionary.
There is probably something else I need to take into account, I'm all ears.