@kerams 2022-04-26
Have you benchmarked typechecking a decently sized file (thousands of lines)?