@nojaf 2022-04-26
This still didn't do all that much in my test.

main branch:

<html>
<body>
<!--StartFragment--><pre style="color: rgb(0, 0, 0); font-style: normal; font-variant-ligatures: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: 2; text-align: start; text-indent: 0px; text-transform: none; widows: 2; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration-thickness: initial; text-decoration-style: initial; text-decoration-color: initial;"><code>BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22000
12th Gen Intel Core i7-12700K, 1 CPU, 20 logical and 12 physical cores
.NET SDK=6.0.202
  [Host]     : .NET 6.0.4 (6.0.422.16404), X64 RyuJIT DEBUG
  DefaultJob : .NET 6.0.4 (6.0.422.16404), X64 RyuJIT
</code></pre><pre style="color: rgb(0, 0, 0); font-style: normal; font-variant-ligatures: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: 2; text-align: start; text-indent: 0px; text-transform: none; widows: 2; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration-thickness: initial; text-decoration-style: initial; text-decoration-color: initial;"><code></code></pre>

Method | Mean | Error | StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated
-- | -- | -- | -- | -- | -- | -- | --
ParseAndCheckFileInProject | 12.11 s | 0.047 s | 0.044 s | 712000.0000 | 180000.0000 | 5000.0000 | 9 GB

<!--EndFragment-->
</body>
</html>

This pr:
<html>
<body>
<!--StartFragment--><pre style="color: rgb(0, 0, 0); font-style: normal; font-variant-ligatures: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: 2; text-align: start; text-indent: 0px; text-transform: none; widows: 2; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration-thickness: initial; text-decoration-style: initial; text-decoration-color: initial;"><code><br class="Apple-interchange-newline">
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22000
12th Gen Intel Core i7-12700K, 1 CPU, 20 logical and 12 physical cores
.NET SDK=6.0.202
  [Host]     : .NET 6.0.4 (6.0.422.16404), X64 RyuJIT DEBUG
  DefaultJob : .NET 6.0.4 (6.0.422.16404), X64 RyuJIT
</code></pre><pre style="color: rgb(0, 0, 0); font-style: normal; font-variant-ligatures: normal; font-variant-caps: normal; font-weight: 400; letter-spacing: normal; orphans: 2; text-align: start; text-indent: 0px; text-transform: none; widows: 2; word-spacing: 0px; -webkit-text-stroke-width: 0px; text-decoration-thickness: initial; text-decoration-style: initial; text-decoration-color: initial;"><code></code></pre>

Method | Mean | Error | StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated
-- | -- | -- | -- | -- | -- | -- | --
ParseAndCheckFileInProject | 11.84 s | 0.035 s | 0.030 s | 711000.0000 | 173000.0000 | 5000.0000 | 9 GB

<!--EndFragment-->
</body>
</html>